Documentation updates for CodeArtifact
This release adds support for UEFI boot on selected AMD- and Intel-based EC2 instances.

Added support to enable AQUA in Amazon Redshift clusters.

This release of the Amazon Macie API adds support for publishing sensitive data findings to AWS Security Hub and specifying which categories of findings to publish to Security Hub.