Added GetCostEstimation and StartCostEstimation to get the monthly resource usage cost and added ability to view resource health by AWS service name and to search insights be AWS service name.
This release adds the ability to search for and order international phone numbers for Amazon Chime SIP media applications.

Documentation updates for health

Enable retrying Training and Tuning Jobs that fail with InternalServerError by setting RetryStrategy.

This release adds the KeyStorageSecurityStandard parameter to the CreateCertificateAuthority API to allow customers to mandate a security standard to which the CA key will be stored within.