Fix to Curl No Proxy Test
In some old platforms the version of Curl used is returning different
error code than the one expected by the tests.
Allowing the tests to pass if the error code received matches
CURLE_SSL_CONNECT_ERROR in addition to the one expected for those
tests that are failing. This is the error code received for example in
Ubuntu 16.04 platform.