Documentation updates for directconnect
Add five new API operations: UpdateGeofenceCollection, UpdateMap, UpdatePlaceIndex, UpdateRouteCalculator, UpdateTracker.

Updated DescribeManagedEndpoint and ListManagedEndpoints to return failureReason and stateDetails in API response.

Documentation updates for reversal of default value for additional instance configuration SSM switch, plus improved descriptions for semantic versioning.

In the Health API, the maximum number of entities for the EventFilter and EntityFilter data types has changed from 100 to 99. This change is related to an internal optimization of the AWS Health service.

This release allows customers to create a new version of WorldTemplates with support for Doors.