AWS DataSync now supports Hadoop Distributed File System (HDFS) Locations
Adds superuser and data-bundle parameters to CreateEnvironment API

This release adds support for specifying the severity of findings that a custom data identifier produces, based on the number of occurrences of text that matches the detection criteria.

This release adds a new boolean attribute - Connect Participant - to the CreateParticipantConnection API, which can be used to mark the participant as connected.